added utils.py for repo-wide funcs
