+++ hera_sim/eor.py
@@ -4,10 +4,11 @@ import numpy as np
 from scipy.interpolate import RectBivariateSpline
 import aipy
 from . import noise
-from . import foregrounds as fg
+from . import utils
 
 
-def noiselike_eor(lsts, fqs, bl_len_ns, eor_amp=1e-5, spec_tilt=0.0, min_dly=0, max_dly=3000):
+def noiselike_eor(lsts, fqs, bl_len_ns, eor_amp=1e-5, spec_tilt=0.0,
+                  fr_width=None, min_dly=0, max_dly=3000):
     """
     Generate a noise-like EoR signal that tracks the sky.
     Modeled after foregrounds.diffuse_foreground().
@@ -19,6 +20,7 @@ def noiselike_eor(lsts, fqs, bl_len_ns, eor_amp=1e-5, spec_tilt=0.0, min_dly=0, 
         eor_amp : float, amplitude of EoR signal [arbitrary]
         spec_tilt : float, spectral slope of EoR spectral amplitude
             as a function of delay in microseconds
+        fr_width : float, width of FR filter in 2pi lambda / sec
         min_dly : float, minimum |delay| in nanosec of EoR signal
         max_dly : float, maximum |delay| in nanosec of EoR signal
 
@@ -26,7 +28,7 @@ def noiselike_eor(lsts, fqs, bl_len_ns, eor_amp=1e-5, spec_tilt=0.0, min_dly=0, 
         vis : 2D ndarray holding simulated complex visibility
     """
     # get fringe rate and generate an LST grid
-    fr_max = np.max(fg.calc_max_fringe_rate(fqs, bl_len_ns))
+    fr_max = np.max(utils.calc_max_fringe_rate(fqs, bl_len_ns))
     dt = 0.5/fr_max # over-resolve by factor of 2
     ntimes = int(np.around(aipy.const.sidereal_day / dt))
     lst_grid = np.linspace(0, 2*np.pi, ntimes, endpoint=False)
@@ -35,7 +37,7 @@ def noiselike_eor(lsts, fqs, bl_len_ns, eor_amp=1e-5, spec_tilt=0.0, min_dly=0, 
     vis = noise.white_noise((ntimes, len(fqs))) * eor_amp
 
     # Fringe-Rate Filter given baseline
-    vis = fg.rough_fringe_filter(vis, lst_grid, fqs, bl_len_ns)
+    vis = utils.rough_fringe_filter(vis, lst_grid, fqs, bl_len_ns, fr_width=fr_width)
 
     # interpolate at fed LSTs
     mdl_real = RectBivariateSpline(lst_grid, fqs, vis.real)
